+++ Source/Widgets/Dojo/CesiumViewerWidget.js
@@ -600,16 +600,21 @@ define([
             reader.readAsText(f);
         },
 
+        _started : false,
+
         /**
          * Call this after placing the widget in the DOM, to initialize the WebGL context,
-         * wire up event callbacks, begin requesting CZML, imagery, etc.  Note this does
-         * not start a render loop (because you may need a custom render loop).
+         * wire up event callbacks, begin requesting CZML, imagery, etc.
          *
          * @function
          * @memberof CesiumViewerWidget.prototype
-         * @see CesiumViewerWidget#startRenderLoop
+         * @see CesiumViewerWidget#autoStartRenderLoop
          */
-        startWidget : function() {
+        startup : function() {
+            if (this._started) {
+                return;
+            }
+
             var canvas = this.canvas, ellipsoid = this.ellipsoid, scene, widget = this;
 
             try {
@@ -620,6 +625,7 @@ define([
                 }
                 return;
             }
+            this._started = true;
 
             this.resize();
 
@@ -893,6 +899,10 @@ define([
             }
 
             this._camera3D = this.scene.getCamera().clone();
+
+            if (this.autoStartRenderLoop) {
+                this.startRenderLoop();
+            }
         },
 
         /**
@@ -1181,13 +1191,23 @@ define([
         },
 
         /**
+         * If true, {@link CesiumViewerWidget#startRenderLoop} will be called automatically
+         * at the end of {@link CesiumViewerWidget#startup}.
+         *
+         * @type {Boolean}
+         * @memberof CesiumViewerWidget.prototype
+         * @default true
+         */
+        autoStartRenderLoop : true,
+
+        /**
          * This is a simple render loop that can be started if there is only one <code>CesiumViewerWidget</code>
          * on your page.  If you wish to customize your render loop, avoid this function and instead
          * use code similar to one of the following examples.
          * @function
          * @memberof CesiumViewerWidget.prototype
          * @see requestAnimationFrame
-         * @see CesiumViewerWidget#startWidget
+         * @see CesiumViewerWidget#autoStartRenderLoop
          * @example
          * // This takes the place of startRenderLoop for a single widget.
          *